[feat] mission add

+++ backend/src/main/java/com/ssafy/edu/model/mission/MissionDifficultRequest.java
@@ -23,5 +23,5 @@ public class MissionDifficultRequest {
 
     @ApiModelProperty(required = true)
     @NotNull
-    int difficulty;
+    double difficulty;
 }
